pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "food-vision"
version = "0.1.0"
description = "Production-grade food image classification with EfficientNet-B2"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.0",
    "torchvision>=0.15",
    "Pillow>=10.0",
    "pyyaml>=6.0",
    "mlflow>=2.10",
    "optuna>=3.5",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "onnx>=1.15",
    "onnxruntime>=1.17",
    "python-multipart>=0.0.9",
    "numpy>=1.26",
    "scikit-learn>=1.4",
    "matplotlib>=3.8",
    "seaborn>=0.13",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "ruff>=0.3",
    "mypy>=1.8",
    "httpx>=0.27",
]

[tool.setuptools.packages.find]
include = ["src*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
